+++ esparto/__init__.py
@@ -2,11 +2,14 @@
 """Top-level package for esparto."""
 
 from importlib.util import find_spec as _find_spec
-from typing import Set
+from pathlib import Path as _Path
+from typing import Set as _Set
 
 __author__ = """Dominic Thorn"""
 __email__ = "dominic.thorn@gmail.com"
-__version__ = "0.2.2"
+__version__ = "0.2.3"
+
+_MODULE_PATH: _Path = _Path(__file__).parent.absolute()
 
 
 _OPTIONAL_DEPENDENCIES: set = {
@@ -16,9 +19,10 @@ _OPTIONAL_DEPENDENCIES: set = {
     "pandas",
     "bokeh",
     "plotly",
+    "weasyprint",
 }
 
-_INSTALLED_MODULES: Set[str] = {
+_INSTALLED_MODULES: _Set[str] = {
     x.name for x in [_find_spec(dep) for dep in _OPTIONAL_DEPENDENCIES] if x
 }
 
@@ -31,3 +35,4 @@ from esparto._content import (
     Markdown,
 )
 from esparto._layout import Column, Page, Row, Section
+from esparto._options import options