+++ tests/test_publish.py
@@ -1,8 +1,10 @@
+from pathlib import Path
 from typing import Optional
 
 import pytest
 from html5lib import HTMLParser  # type: ignore
 
+import esparto as es
 import esparto._publish as pu
 from tests.conftest import _EXTRAS, content_list, layout_list
 
@@ -28,22 +30,49 @@ def test_content_html_valid(content):
     assert html_is_valid(html, fragment=True)
 
 
-def test_rendered_html_valid(page_layout, tmp_path):
+def test_rendered_html_valid_cdn(page_layout: es.Page, tmp_path):
     path = str(tmp_path / "my_page.html")
-    html = pu.publish(page_layout, path, return_html=True)
+    html = pu.publish_html(page_layout, path, return_html=True, dependency_source="cdn")
     assert html_is_valid(html)
 
 
-def test_saved_html_valid(page_layout, tmp_path):
+def test_rendered_html_valid_inline(page_layout: es.Page, tmp_path):
     path = str(tmp_path / "my_page.html")
-    page_layout.save(path)
-    with open(path, "r") as f:
-        html = f.read()
+    html = pu.publish_html(
+        page_layout, path, return_html=True, dependency_source="inline"
+    )
+    assert html_is_valid(html)
+
+
+def test_saved_html_valid_cdn(page_layout: es.Page, tmp_path):
+    path: Path = tmp_path / "my_page.html"
+    page_layout.save_html(str(path), dependency_source="cdn")
+    html = path.read_text()
+    assert html_is_valid(html)
+
+
+def test_saved_html_valid_inline(page_layout: es.Page, tmp_path):
+    path: Path = tmp_path / "my_page.html"
+    page_layout.save_html(str(path), dependency_source="inline")
+    html = path.read_text()
     assert html_is_valid(html)
 
 
 if _EXTRAS:
 
-    def test_notebook_html_valid(page_layout):
-        html = pu.nb_display(page_layout, return_html=True)
+    def test_notebook_html_valid_cdn(page_layout):
+        html = pu.nb_display(page_layout, return_html=True, dependency_source="cdn")
         assert html_is_valid(html)
+
+    def test_notebook_html_valid_inline(page_layout):
+        html = pu.nb_display(page_layout, return_html=True, dependency_source="inline")
+        assert html_is_valid(html)
+
+    @pytest.mark.parametrize("content", content_list)
+    def test_pdf_output(content, tmp_path):
+        if "bokeh" not in content._dependencies:
+            page = es.Page(content)
+            path: Path = tmp_path / "my_page.pdf"
+            page.save_pdf(str(path))
+            size = path.stat().st_size
+            assert size > 1000